+++ examples/basic/purplet.config.js
@@ -1,0 +1,6 @@
+/** @type {import('purplet').Config} */
+const config = {
+
+};
+
+export default config;