+++ packages/purplet/src/config/default.ts
@@ -1,15 +1,22 @@
-import type { ResolvedConfig } from './options';
+// This file is used to populate the default configuration in the documentation.
+import type { ResolvedConfig } from './types';
 
 const config: ResolvedConfig = {
+  root: '',
+  // START CONFIG
   alias: {
+    $features: 'src/features',
     $lib: 'src/lib',
   },
+  lang: 'en-US',
   paths: {
+    build: 'dist',
     features: 'src/features',
-    output: '.purplet',
-    translations: './todo/undetermined',
+    temp: '.purplet',
+    translations: 'lang',
   },
   vite: () => ({}),
+  // END CONFIG
 };
 
 export default config;