+++ packages/purplet/src/hooks/command.ts
@@ -1,5 +1,5 @@
-import { ApplicationCommandType } from 'discord-api-types/v10';
-import type { Interaction } from 'discord.js';
+import { ApplicationCommandOptionType, ApplicationCommandType } from 'discord-api-types/v10';
+import { CommandInteraction, Interaction } from 'discord.js';
 import { ApplicationCommandData, createFeature } from '../lib/feature';
 
 function formatCommandName(cmd: ApplicationCommandData) {
@@ -18,15 +18,31 @@ export interface ApplicationCommandHookData {
   handle(this: Interaction): void;
 }
 
+export function getFullCommandName(interaction: any) {
+  return interaction.commandType === ApplicationCommandType.ChatInput
+    ? [
+        interaction.commandName,
+        interaction.options.data.find(
+          (x: any) => x.type === ApplicationCommandOptionType.SubcommandGroup
+        )?.name,
+        interaction.options.data.find(
+          (x: any) => x.type === ApplicationCommandOptionType.Subcommand
+        )?.name,
+      ]
+        .filter(Boolean)
+        .join(' ')
+    : interaction.commandName;
+}
+
 export function $applicationCommand(opts: ApplicationCommandHookData) {
   return createFeature({
     name: `command: ${formatCommandName(opts.command)}`,
     applicationCommands: [opts.command],
     interaction(i) {
       if (
-        i.isCommand() &&
+        i instanceof CommandInteraction &&
         i.commandType === opts.command.type &&
-        i.commandName === opts.command.name
+        getFullCommandName(i) === opts.command.name
       ) {
         opts.handle.call(i);
       }