+++ packages/purplet/src/index.ts
@@ -1,7 +1,7 @@
 // `purplet` public api entry point:
 
 export * from './builders';
-export type { Config } from './config/options';
+export type { Config } from './config/types';
 export * from './hooks';
 export * from './lib/feature';
 export * from './lib/gateway';